Split out getPlanRawDiscount
